1st Finish

+++ include/agimus_vision/tracker_object/fwd.hpp
@@ -21,7 +21,7 @@ typedef vpImage< unsigned char > GrayImage_t;
 typedef vpImage< vpRGBa > RGBaImage_t;
 
 // depth map image
-typedef vpImage<float> DepthMap_t;
+typedef vpImage<u_int16_t> DepthMap_t;
 
 struct DetectorAprilTagWrapper;
 typedef std::shared_ptr<DetectorAprilTagWrapper> DetectorAprilTagWrapperPtr;
